Lots of work (#48)
* Added initial overlay refactor


* Added test coverage for complex nested overlays

* Added auto dismiss capable toast implementation


* Added useLvglVisibility hook


* checkpoint


* checkpoint


* checkpoint


* checkpioint


* folder reorg

* folder reorg 2


* Removed yaml logic from core


* checkpoint


* checkpoint


* Adde HAEntityVariant


* checkpoint


* checkpoint


* checkpoint


* Fixed rotation issue


* Source mode compliler working

* Cleanup


* Semantic html viewer working-ish


* Moved viewer file to better location


* Added kind and typed IR nodes for registries


* Cleaned up IR tree

* Simplified cpp with a helper function

* Removed old branding


* Added changeset notes


---------

+++ packages/cli/scripts/prebuild.js
@@ -13,10 +13,16 @@ const assets = path.resolve(__dirname, '..', 'assets');
 
 const copies = [
   { src: '../../esphome-target/assets/external-component', dest: 'external-component' },
+  { src: '../src/assets/ir-viewer.html', dest: 'ir-viewer.html', file: true },
 ];
 
-for (const { src, dest } of copies) {
+for (const { src, dest, file } of copies) {
   const target = path.join(assets, dest);
-  fs.rmSync(target, { recursive: true, force: true });
-  fs.cpSync(path.resolve(__dirname, src), target, { recursive: true });
+  if (file) {
+    fs.mkdirSync(path.dirname(target), { recursive: true });
+    fs.copyFileSync(path.resolve(__dirname, src), target);
+  } else {
+    fs.rmSync(target, { recursive: true, force: true });
+    fs.cpSync(path.resolve(__dirname, src), target, { recursive: true });
+  }
 }
